Version 2.0-dev
This is a major rework of the entire script.

NOTICE: YOU WILL NEED TO REDO YOUR CONFIG, AS THE PREVIOUS ONE IS NOT COMPATIBLE WITH THIS!

UI has changed their entire OS on the controller, and thus some changes were done with SSL certs.
Since there is a large number of fariables at play now, it was found best to rework this entirely to be more dynamic and capable of applying to more deployments.

Some effort will be required from users self hosting their controller, whether it's on-premise or in the cloud. You will need to locate the destination paths and set the configs manually.

This is currently completely untested. It will remain in the dev branch until it has been tested and validated. This push is simply the first view of the major workover, and a checkpoint for it.